app/requirements.txt
pandas==1.5.3
python-dotenv==1.0.0
openai==0.27.0
azure-identity==1.12.0
pdf2image==1.16.3
IPython==8.12.0
Pillow==9.4.0